+++ tegral-di/tegral-di-core/src/main/kotlin/guru/zoroark/tegral/di/dsl/ContextBuilderDsl.kt
@@ -156,7 +156,7 @@ fun <T : Any> ContextBuilderDsl.put(kclass: KClass<T>, qualifier: Qualifier, sup
         supplier.parameters.size == 1 && supplier.parameters.first().type == InjectionScope::class.createType() ->
             put(kclass, qualifier) { supplier.call(scope) }
         else -> throw InvalidDeclarationException(
-            "Cannot 'put' the given function ($supplier). It must take either no arguments or a single argument " +
+            "Cannot 'put' the given function ($supplier). It must take either no parameters or a single argument " +
                 "that is of type 'InjectionScope'. Consider manually instantiating this component instead."
         )
     }